requirements.txt
akshare
baostock
cn-stats
pandas
PyYAML
pypdf
pywencai
requests
